Implement ISO 9564-1:2017 PIN block format 4
This format requires separate PIN field and PAN field encoding functions
because encryption occurs before the fields are combined. It is
therefore the caller's responsibility to encrypt and combine these
encoded fields in accordance with ISO 9564-1:2017 9.4.2.3. Similarly,
the caller must separate and decrypt PIN field before it can be decoded.